simplify signing
